backend/transcription/asr_gpu/requirements.txt
python-dotenv
numpy
soundfile
librosa
ffmpeg-python
faster-whisper
whisperx
ctranslate2
pyannote.audio
psycopg2-binary
requests
tqdm
